requirements.txt
streamlit>=1.10.0
spacy>=3.0.0
pdfminer.six
python-docx
sentence-transformers
